Initial work to create a UnboundInput component that is decoupled from the form data.
