+++ src/index.js
@@ -1,5 +1,6 @@
 import Form from "./components/form.js"
 import Input from "./components/input.js"
+import UnboundInput from "./components/unbound_input.js"
 import Submit from "./components/submit.js"
 import FormErrorList from "./components/form_error_list.js"
 import Fieldset from "./components/fieldset.js"
@@ -26,6 +27,7 @@ export default {
   defaultTheme,
   Form,
   Input,
+  UnboundInput,
   Submit,
   FormErrorList,
   Fieldset,
